% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_cos_sim_bars}
\alias{plot_cos_sim_bars}
\title{Plot cosine similarity bars}
\usage{
plot_cos_sim_bars(res_dt)
}
\arguments{
\item{res_dt}{a result from \code{\link{stransevalr()}}}
}
\description{
Plot cosine similarity bars
}
