% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_cos_sim_boxes}
\alias{plot_cos_sim_boxes}
\title{Plot cosine similarity box plots by model}
\usage{
plot_cos_sim_boxes(res_dt)
}
\arguments{
\item{res_dt}{a result from \code{\link{stransevalr()}}}
}
\description{
Plot cosine similarity box plots by model
}
