% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stransevalr-package.R
\docType{package}
\name{stransevalr-package}
\alias{stransevalr-package}
\title{stransevalr: What the Package Does (Title Case)}
\description{
Use sentence transformers to evaluate model responses in R.
}
\author{
\strong{Maintainer}: Andrew Ghazi \email{andrew_ghazi@hms.harvard.edu}

}
\keyword{internal}
