% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stransevalr.R
\name{stransevalr}
\alias{stransevalr}
\title{Evaluate model responses against a "ground-truth" answer}
\usage{
stransevalr(input, sent_trans_model = "multi-qa-MiniLM-L6-cos-v1")
}
\arguments{
\item{input}{an input data frame or path to tsv}

\item{sent_trans_model}{hugging face sentence transformer model string}
}
\description{
Starting with an input set of questions, answers, and model
responses, use sentence transformers to compare the embedded model responses
against the embedded "ground-truth" answer.
}
\details{
If \code{sent_trans_model} has not been run before, it should download it from huggingface as needed.
}
